refactor(qbittorrent-e2e): extract compose stack provisioning into compose_stack module
